Week 3/Day 1/legal-summarizer-app/requirements.txt
chainlit==0.7.700
transformers==4.41.2
peft==0.11.1
tiktoken==0.5.1
python-dotenv==1.0.0